User Shift Update

+++ database/migrations/2022_07_26_123035_create_export_pack_sizes_table.php
@@ -16,7 +16,7 @@ class CreateExportPackSizesTable extends Migration
         Schema::create('export_pack_sizes', function (Blueprint $table) {
             $table->id();
             $table->string('name')->nullable();
-            $table->integer('weight')->nullable();
+            $table->double('weight')->nullable();
             $table->timestamps();
         });
     }
